+++ part-1-batched-adsorption/scripts/oc20dense_dft_reference_checks.py
@@ -57,7 +57,7 @@ DEFAULT_TOOLKIT_ROOT = (
     / "outputs"
     / "precomputed"
     / "accuracy"
-    / "oc20dense_closed_shell_trajectory_mh1_oc20_usemppbe"
+    / "oc20dense_closed_shell_trajectory_mace_mpa0"
 )
 DEFAULT_ARCHIVE = DEFAULT_TRAJECTORY_ARCHIVE
 DEFAULT_OUTDIR = DEFAULT_TOOLKIT_ROOT / "dft_reference_checks"